@@ -1,6 +1,6 @@
 {
   "name": "admin-one-vue-bulma-dashboard",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
@@ -9,37 +9,34 @@
   },
   "dependencies": {
     "@mdi/js": "^6.5.95",
-    "@vue/composition-api": "^1.4.1",
+    "@oruga-ui/oruga-next": "^0.4.8",
+    "@oruga-ui/theme-bulma": "^0.1.5",
     "axios": "^0.24.0",
-    "buefy": "^0.9.13",
     "bulma": "^0.9.3",
-    "chart.js": "^2.9.4",
+    "chart.js": "^3.6.1",
     "core-js": "^3.19.1",
     "lodash": "^4.17.20",
     "numeral": "^2.0.6",
-    "vue": "^2.6.12",
-    "vue-chartjs": "^3.5.1",
-    "vue-router": "^3.5.3",
-    "vuex": "^3.6.2"
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0-0",
+    "vuex": "^4.0.0-0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.11",
-    "@vue/cli-plugin-eslint": "^4.5.11",
-    "@vue/cli-plugin-router": "^4.5.11",
-    "@vue/cli-plugin-vuex": "^4.5.11",
-    "@vue/cli-service": "^4.5.11",
-    "@vue/eslint-config-standard": "^5.1.2",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.8.0",
+    "@babel/eslint-parser": "^7.16.3",
+    "@vue/cli-plugin-babel": "5.0.0-rc.1",
+    "@vue/cli-plugin-eslint": "5.0.0-rc.1",
+    "@vue/cli-plugin-router": "5.0.0-rc.1",
+    "@vue/cli-plugin-vuex": "5.0.0-rc.1",
+    "@vue/cli-service": "5.0.0-rc.1",
+    "@vue/compiler-sfc": "^3.0.0",
+    "@vue/eslint-config-standard": "^6.0.0",
+    "eslint": "^7.32.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.3.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "glob-parent": "^6.0.2",
+    "eslint-plugin-vue": "^7.20.0",
     "sass": "^1.32.7",
-    "sass-loader": "^10.1.1",
-    "vue-template-compiler": "^2.6.12"
+    "sass-loader": "^10.1.1"
   },
   "browserslist": [
     "> 1%",
