@@ -3,12 +3,17 @@ module.exports = {
   env: {
     node: true
   },
+  globals: {
+    defineProps: 'readonly',
+    defineEmits: 'readonly'
+  },
   extends: [
-    'plugin:vue/recommended',
+    'plugin:vue/vue3-recommended',
     '@vue/standard'
   ],
+  parser: 'vue-eslint-parser',
   parserOptions: {
-    parser: 'babel-eslint'
+    parser: '@babel/eslint-parser'
   },
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
