@@ -40,7 +40,7 @@
 </template>
 
 <script>
-import { computed, ref } from '@vue/composition-api'
+import { computed, ref } from 'vue'
 
 export default {
   name: 'AsideMenuItem',
