@@ -21,7 +21,7 @@
 </template>
 
 <script>
-import { computed } from '@vue/composition-api'
+import { computed } from 'vue'
 
 export default {
   name: 'CheckboxRadioPicker',
