@@ -1,4 +1,5 @@
 <script>
+import { h } from 'vue'
 import chunk from 'lodash/chunk'
 
 export default {
@@ -9,24 +10,26 @@ export default {
       default: 5
     }
   },
-  render (createElement) {
-    const renderAncestor = elements => createElement(
+  render () {
+    const renderAncestor = elements => h(
       'div',
       { attrs: { class: 'tile is-ancestor' } },
       elements.map((element) => {
-        return createElement('div', { attrs: { class: 'tile is-parent' } }, [
+        return h('div', { attrs: { class: 'tile is-parent' } }, [
           element
         ])
       })
     )
 
-    if (this.$slots.default.length <= this.maxPerRow) {
-      return renderAncestor(this.$slots.default)
+    // const slots = useSlots()
+
+    if (this.$slots.default().length <= this.maxPerRow) {
+      return renderAncestor(this.$slots.default())
     } else {
-      return createElement(
+      return h(
         'div',
         { attrs: { class: 'is-tiles-wrapper' } },
-        chunk(this.$slots.default, this.maxPerRow).map((group) => {
+        chunk(this.$slots.default(), this.maxPerRow).map((group) => {
           return renderAncestor(group)
         })
       )
